Steps to creating a project on a local machine
This commit contains the initial steps to creating a project on a local machine and setting up a Git repository.  This way, when I go to set up this project on a different machine I don't have to research everything again.  Or if I invite someone to help me they can follow the instructions.